created Permission.vue view; createt getPermissions,renamePermission,deletePermission methods; created roles() relationship in Permission model; created PermissionResource; upd api.php routes;

+++ app/Http/Controllers/Admin/Permission/PermissionController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Admin\Permission;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Admin\Permission\PermissionRequest;
+use App\Http\Resources\Admin\Permission\PermissionResource;
 use App\Http\Resources\Admin\Role\RolePermissionResource;
 use App\Models\Permission;
 
@@ -13,14 +14,19 @@ class PermissionController extends Controller
     public function index()
     {
         $permissions = Permission::all();
+        return response()->json(['permissions' => PermissionResource::collection($permissions)]);
+    }
+
+    public function getPermissionsForForm()
+    {
+        $permissions = Permission::all();
         return response()->json(['permissions' => RolePermissionResource::collection($permissions)]);
     }
 
     public function store(PermissionRequest $request)
     {
         $data = $request->validated();
-        $data['slug'] = str_replace(' ', '-', strtolower($data['name']));
-        //dd($data);
+        $data['slug'] = 'can-' . str_replace(' ', '-', strtolower($data['name']));
         $permission = Permission::firstOrCreate($data);
         return response()->json([
             'message' => 'Permission created!',
@@ -36,9 +42,8 @@ class PermissionController extends Controller
     public function update(PermissionRequest $request, Permission $permission)
     {
         $data = $request->validated();
-        $data['slug'] = str_replace(' ', '-', strtolower($data['name']));
-        //dd($data);
-        foreach ($data as $key => $value){
+        $data['slug'] =  "can-" . str_replace(' ', '-', strtolower($data['name']));
+        foreach ($data as $key => $value) {
             $permission->$key = $value;
         }
         $permission->save();
@@ -50,6 +55,9 @@ class PermissionController extends Controller
 
     public function delete(Permission $permission)
     {
+        if(count($permission->roles)!==0){
+            return response()->json(['message' => 'This permission is linked to one of the roles. Remove it from the role permissions list'], 422);
+        }
         $permission->delete();
         return response()->json(['message' => 'Permission deleted successfully!']);
     }
